+++ packages/core/transpiler/view-generator/src/NodeGenerators/ForGenerator.ts
@@ -13,7 +13,14 @@ export default class ForGenerator extends BaseGenerator {
 
     // ---- Declare for node
     this.addInitStatement(
-      this.declareForNode(dlNodeName, array.value, item, children, key)
+      this.declareForNode(
+        dlNodeName,
+        array.value,
+        item,
+        children,
+        BaseGenerator.calcDependencyNum(array.dependencyIndexArr),
+        key
+      )
     )
 
     // ---- Update statements
@@ -38,13 +45,14 @@ export default class ForGenerator extends BaseGenerator {
    *   }
    *   ${topLevelNodes[0])._$updateFunc = $update
    *   return [...${topLevelNodes}]
-   * }, ${array}.map(${item} => ${key}))
+   * }, ${depNum}, ${array}.map(${item} => ${key}))
    */
   private declareForNode(
     dlNodeName: string,
     array: t.Expression,
     item: t.LVal,
     children: ViewParticle[],
+    depNum: number,
     key?: t.Expression
   ): t.VariableDeclaration {
     // ---- NodeFunc
@@ -82,6 +90,7 @@ export default class ForGenerator extends BaseGenerator {
             [item as any],
             this.t.blockStatement(childStatements)
           ),
+          this.t.numericLiteral(depNum),
           ...this.getForKeyStatement(array, item, key),
         ])
       ),