+++ packages/core/transpiler/view-generator/src/NodeGenerators/IfGenerator.ts
@@ -4,14 +4,20 @@ import { type IfParticle, type IfBranch } from "@dlightjs/reactivity-parser"
 
 export default class IfGenerator extends BaseGenerator {
   run() {
-    const ifParticle = this.viewParticle as IfParticle
-    // ---- declareIfNode
-    const dlNodeName = this.generateNodeName()
-    this.addInitStatement(this.declareIfNode(dlNodeName, ifParticle.branches))
-
-    const deps = ifParticle.branches.flatMap(
+    const { branches } = this.viewParticle as IfParticle
+    const deps = branches.flatMap(
       ({ condition }) => condition.dependencyIndexArr ?? []
     )
+    // ---- declareIfNode
+    const dlNodeName = this.generateNodeName()
+    this.addInitStatement(
+      this.declareIfNode(
+        dlNodeName,
+        branches,
+        BaseGenerator.calcDependencyNum(deps)
+      )
+    )
+
     this.addUpdateStatements(deps, this.updateIfNodeCond(dlNodeName))
     this.addUpdateStatementsWithoutDep(this.updateIfNode(dlNodeName))
 
@@ -103,9 +109,13 @@ export default class IfGenerator extends BaseGenerator {
    *    thisIf.cond = 1
    *    return [nodes]
    *   }
-   * })
+   * }, ${depNum})
    */
-  private declareIfNode(dlNodeName: string, branches: IfBranch[]): t.Statement {
+  private declareIfNode(
+    dlNodeName: string,
+    branches: IfBranch[],
+    depNum: number
+  ): t.Statement {
     let hasElseStatement = false
 
     const ifStatement = branches
@@ -166,6 +176,7 @@ export default class IfGenerator extends BaseGenerator {
             [this.t.identifier("$thisIf")],
             this.t.blockStatement([ifStatement])
           ),
+          this.t.numericLiteral(depNum),
         ])
       ),
     ])