+++ packages/core/dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/dlight
 
+## 1.0.0-alpha.3
+
+### Patch Changes
+
+- fix: no redundant update for new nodes in if and for
+
 ## 1.0.0-alpha.2
 
 ### Patch Changes